♻️ [Refactor] NoteViewModel 삭제
- 뷰모델의 Observable 타입 즉, 관찰하는 타입을 Note로 수정
- 날짜 포맷팅 기능을 NoteListViewModel에 구현함
